Medicine list fixes
